refactor: apply simplification review
Shared hold-payload/note/provenance helpers exported and reused;
provider inheritance hoisted with the fallback deferred to the config
seam; listLocalTemplates uses parsePluginManifest; dead timeoutMs
option and unread outcome field removed; template-list walk prunes at
each plugin.json; template test scaffold folded into create-agent.test;
tool/verb descriptions made mechanical; docs skew note made generic.
Net -119 lines.
